# public/spec/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API service used for handwriting transcriptions",
    "version": "0.0.0",
    "title": "Moleskine ICR API",
    "contact": {
      "email": "gabriel.walsh@moleskine.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "msk-icr.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/transcribe":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Transcribes a photographed Moleskine Page.",
        "consumes": [
          "application/json",
          "x-www-form-urlencoded"
        ],
        "description": "Takes the request for a given image URI and provider and provides a transcription.",
        "operationId": "getTranscription",
        "parameters": [
          {
            "in": "body",
            "name": "imageUri",
            "description": "The URI of the image to be transcribed.",
            "schema": {
              "$ref": "#/definitions/Image"
            }
          },
          {
            "in": "body",
            "name": "imageFile",
            "description": "JPEG needing to be transcribed",
            "type": "binary"
          }
        ],
        "produces":[
          "application/json"
        ],
        "responses":{
          "201":{
            "description": "Successful transcription",
            "schema":{
              "type":"array",
              "items": {
                "$ref": "#/definitions/Transcription"
              }
            }
          },
          "500":{
            "description": "Server error"
          }
        }
      },
      "security": [
        {
          "user_auth": [
            "read:providers"
          ]
        }
      ]
    },
    "/image":{
      "get":{
        "tags": [
          "image"
        ],
        "operationId": "getImage",
        "summary": "Retrieves an image given an id.",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"query",
            "required": "true",
            "type": "string"
          }
        ],
        "responses":{
          "200":{
            "description": "Successful retrieval",
            "schema":{
                "$ref": "#/definitions/Image"
            }
          },
          "500":{
            "description": "Error"
          }
        }
      },
      "security": [
        {
          "user_auth": [
            "read:providers"
          ]
        }
      ]
    },
    "/image/create":{
      "post":{
        "tags": [
          "image"
        ],
        "summary": "Create Image",
        "description": "Create an image record in the db",
        "operationId": "createImage",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name":"uri",
            "in":"query",
            "required": "true",
            "type": "string"
          }
        ],
        "responses":{
          "200":{
            "description": "Successful creation",
            "schema":{
              "$ref": "#/definitions/Image"
            }
          },
          "500":{
            "description": "Error"
          }
        },
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/image/delete":{
      "post":{
        "tags": [
          "image"
        ],
        "summary": "Delete Image",
        "description": "Deletes an image record in the db",
        "operationId": "deleteImage",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"query",
            "required": "true",
            "type": "string"
          }
        ],
        "responses":{
          "200":{
            "description": "Successful Deletion"
          },
          "500":{
            "description": "Error"
          }
        },
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/image/update":{
      "post":{
        "tags": [
          "image"
        ],
        "summary": "Update Image",
        "description": "Updates an image record in the db",
        "operationId": "updateImage",
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "name":"id",
            "in":"query",
            "required": "true",
            "type": "string"
          },
          {
            "name":"uri",
            "in":"query",
            "required": "false",
            "type": "string"
          }
        ],
        "responses":{
          "200":{
            "description": "Successful update",
            "schema":{
              "$ref": "#/definitions/Image"
            }
          },
          "500":{
            "description": "Error"
          }
        },
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/delete-temp":{
      "post":{
        "tags": [
          "image",
          "file"
        ],
        "summary": "Delete Temp",
        "description": "Deletes a temporary upload file in the /tmp directory",
        "operationId": "deleteImage",
        "responses":{
          "200":{
            "description": "Successful deletion"
          },
          "500":{
            "description": "Error"
          }
        },
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/providers":{
      "get": {
        "tags": [
          "provider"
        ],
        "summary": "List Providers",
        "description": "Shows the current ICR and NLP providers.",
        "operationId": "getProviders",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful listing",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Provider"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          },
          "500": {
            "description": "Server Error"
          }
        },
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/providers/seed":{
      "post":{
        "tags": [
          "provider"
        ],
        "summary": "Seed Providers",
        "description": "Seeds provides from preset JSON..",
        "operationId": "seedProviders",
        "produces": [
          "application/json"
        ],
        "responses":{
          "200":{
            "description": "Successful seeding",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Provider"
              }
            }
          },
          "500":{
            "description": "Error"
          }
        },
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/provider/":{
      "post":{
        "tags": [
          "provider"
        ],
        "summary": "Get Provider",
        "description": "Retrieves a provider given a specific id.",
        "operationId": "getProvider",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/provider/create":{
      "post":{
        "tags": [
          "provider"
        ],
        "summary": "Create Provider",
        "description": "Creates a new Provider",
        "operationId": "createProvider",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/provider/delete":{
      "post":{
        "tags": [
          "provider"
        ],
        "summary": "Delete Provider",
        "description": "Deletes a new Provider",
        "operationId": "deleteProvider",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/provider/update":{
      "post":{
        "tags": [
          "provider"
        ],
        "summary": "Update Provider",
        "description": "Updates a given Provider",
        "operationId": "updateProvider",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/transcription":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Get Transcription",
        "description": "Gets a given Transcription",
        "operationId": "getTranscription",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/transcription/create":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Create Transcription",
        "description": "Creates a newTranscription",
        "operationId": "createTranscription",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/transcription/delete":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Delete Transcription",
        "description": "Deletes a given Transcription",
        "operationId": "deleteTranscription",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/transcription/update":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Update Transcription",
        "description": "Updates a given Transcription",
        "operationId": "updateTranscription",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/user":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Get User",
        "description": "Gets a given user",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/user/create":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Create User",
        "description": "Creates a new user",
        "operationId": "createUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/user/delete":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Delete User",
        "description": "Deletes a given user",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    },
    "/user/update":{
      "post":{
        "tags": [
          "transcription"
        ],
        "summary": "Update User",
        "description": "Updates a given user",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "user_auth": [
              "read:providers"
            ]
          }
        ]
      }
    }
  },
  "securityDefinitions": {
    "user_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://msk-icr.herokuapp.com/oauth/providers",
      "flow": "implicit",
      "scopes": {
        "read:providers": "lists providers"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "Image": {
      "type": "object",
      "properties": {
        "uri": {
          "type": "string",
          "description": "URI of the photographed Notebook page"
        },
        "id":{
          "type": "string",
          "description": "Id of the image record."
        }
      }
    },
    "Provider": {
      "type": "object",
      "properties": {
        "apiId": {
          "type": "string",
          "description": "API id for each provider passed during transcription."
        },
        "name": {
          "type": "string",
          "description": "Name of ICR or NLP Provider"
        },
        "description": {
          "type": "string",
          "description": "Description of ICR or NLP Provider"
        },
        "documentation": {
          "type": "string",
          "format": "URL to further documentation"
        },
        "active": {
          "type": "boolean",
          "description": "Whether provider is currently active."
        }
      }
    },
    "Transcription":{
      "type": "object",
      "properties": {
        "provider": {
          "type": "string",
          "description": {
            "$ref": "#/definitions/Provider"
          }
        },
        "id":{
          "type":"string",
          "description": "ID of transcription object."
        },
        "text":{
          "type":"string",
          "description": "Text transcribed from image"
        }
      }
    }
  }
}